// src/app/experience/interfaces.tsx
export interface Experience {
    position: string;
    company: string;
    period: string;
    location: string;
    activities: string[];
    website?: string;
    logo?: string;
    currentJob?: boolean;
    skills?: string[];
    duration: string;
  }
  

export interface DuracaoTotal {
    anos: number;
    meses: number;
  }
  